refactor(frontend): extra machine service list into a composable
Extract logic to get list of machine services into a useMachineServices composable.
